get fqdn from /etc/hosts for linux and darwin
