+++ providers/linux/host_linux.go
@@ -18,14 +18,11 @@
 package linux
 
 import (
-	"bufio"
 	"errors"
 	"fmt"
-	"io/fs"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/joeshaw/multierror"
@@ -152,7 +149,6 @@ func newHost(fs procFS) (*host, error) {
 	r.bootTime(h)
 	r.containerized(h)
 	r.hostname(h)
-	r.domain(h)
 	r.fqdn(h)
 	r.network(h)
 	r.kernelVersion(h)
@@ -216,40 +212,13 @@ func (r *reader) hostname(h *host) {
 	h.info.Hostname = v
 }
 
-func (r *reader) domain(h *host) {
-	v, err := domainname()
+func (r *reader) fqdn(h *host) {
+	v, err := shared.FQDN()
 	if r.addErr(err) {
 		return
 	}
-	h.info.Domain = v
-}
 
-const etcHosts = "/etc/hosts"
-
-func (r *reader) fqdn(h *host) {
-	f, err := os.Open(etcHosts)
-	if err != nil {
-		r.addErr(fmt.Errorf("could open %q to get FQDN: %w", etcHosts, err))
-		return
-	}
-
-	hname, err := os.Hostname()
-	if err != nil {
-		r.addErr(fmt.Errorf("could get hostname to look for FQDN: %w", err))
-		return
-	}
-
-	fqdn, err := fqdnFromHosts(hname, f)
-	if err != nil {
-		r.addErr(fmt.Errorf("error when looking for FQDN on %s: %w", etcHosts, err))
-		return
-	}
-
-	if fqdn == "" {
-		// FQDN not found on hosts file, fall back to net.Lookup?
-		// add an error?
-	}
-	h.info.FQDN = fqdn
+	h.info.FQDN = v
 }
 
 func (r *reader) network(h *host) {
@@ -297,69 +266,4 @@ type procFS struct {
 func (fs *procFS) path(p ...string) string {
 	elem := append([]string{fs.mountPoint}, p...)
 	return filepath.Join(elem...)
-}
-
-// fqdnFromHosts looks for the FQDN for hostname on hostFile.
-// If successfully it returns FQDN, nil. If no FQDN for hostname is found
-// it returns "", nil. It returns "", err if any error happens.
-func fqdnFromHosts(hostname string, hostsFile fs.File) (string, error) {
-	s := bufio.NewScanner(hostsFile)
-
-	for s.Scan() {
-		fqdn := findInHostsLine(hostname, s.Text())
-		if fqdn != "" {
-			return fqdn, nil
-		}
-	}
-	if err := s.Err(); err != nil {
-		return "", fmt.Errorf("error reading hosts file lines: %w", err)
-	}
-
-	return "", nil
-}
-
-// findInHostsLine takes a HOSTS(5) line and searches for an alias matching
-// hostname, if found it returns the canonical_hostname. The canonical_hostname
-// should be the FQDN, see HOSTNAME(1).
-// TODO: check k8s: https://kubernetes.io/docs/tasks/network/customize-hosts-file-for-pods/
-func findInHostsLine(hostname, hostsEntry string) string {
-	line, _, _ := strings.Cut(hostsEntry, "#")
-	if len(line) < 1 {
-		fmt.Printf("skip comment or empty: %q\n", hostsEntry)
-		return ""
-	}
-
-	fileds := strings.FieldsFunc(line, func(r rune) bool {
-		return r == ' ' || r == '\t'
-	})
-
-	if len(fileds) < 2 {
-		// invalid hostsEntry
-		return ""
-	}
-
-	// fields[0] is the ip address
-	cannonical, aliases := fileds[1], fileds[1:]
-
-	// TODO: confirm: a name should not repeat on different addresses.
-	if len(fileds) == 2 {
-		if fileds[1] == hostname {
-			return cannonical
-		}
-
-		// If hostname was not set as an alias for FQDN, but the fist name
-		// before the dot is the hostname:
-		//   192.168.1.10    foo.mydomain.org	#  foo
-		if hname, _, _ := strings.Cut(cannonical, "."); hname == hostname {
-			return cannonical
-		}
-	}
-
-	for _, h := range aliases {
-		if h == hostname {
-			return cannonical
-		}
-	}
-
-	return ""
 }
