+++ providers/linux/host_linux_test.go
@@ -109,64 +109,3 @@ func TestHostNetworkCounters(t *testing.T) {
 	}
 	t.Log(string(data))
 }
-
-func TestParseLine(t *testing.T) {
-	tcs := []struct {
-		name     string
-		hostname string
-		line     string
-		want     string
-	}{
-		{
-			name:     "find fqdn - spaces",
-			hostname: "thishost",
-			want:     "thishost.mydomain.org",
-			line:     "127.0.1.1       thishost.mydomain.org  thishost",
-		},
-		{
-			name:     "find fqdn - tabs",
-			hostname: "thishost",
-			want:     "thishost.mydomain.org",
-			line:     "127.0.1.1	thishost.mydomain.org	thishost",
-		},
-		{
-			name:     "find fqdn - tabs and spaces",
-			hostname: "thishost",
-			want:     "thishost.mydomain.org",
-			line:     "127.0.1.1	 thishost.mydomain.org	  thishost",
-		},
-		{
-			name:     "find fqdn - line with comment",
-			hostname: "bar",
-			want:     "bar.mydomain.org",
-			line:     "192.168.1.13    bar.mydomain.org       bar # comment 1",
-		},
-		{
-			name:     "find fqdn - fqdn with hostname but no alias",
-			hostname: "ahostWith",
-			want:     "ahostWith.no.alias",
-			line:     "213.456.178.9 ahostWith.no.alias",
-		},
-		{
-			name:     "ignore invalid line",
-			hostname: "ignore",
-			want:     "",
-			line:     "209.237.226.91INVALIDwww.opensource.org",
-		},
-		{
-			name:     "comment line",
-			hostname: "ignore",
-			want:     "",
-			line:     "# The following lines are desirable for IPv4 capable hosts",
-		},
-	}
-
-	for _, tc := range tcs {
-		t.Run(tc.name, func(t *testing.T) {
-			got := findInHostsLine(tc.hostname, tc.line)
-			if got != tc.want {
-				t.Errorf("got %s, want %s", got, tc.want)
-			}
-		})
-	}
-}
